backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
requests==2.31.0
django-filter==23.4
python-dotenv==1.0.0
django-cors-headers==3.14.0
loguru==0.6.0
mysqlclient==2.1.1
django-cors-headers==3.14.0
django-redis==5.4.0
hiredis==2.3.2
redis==5.0.1
